Event management
Player can iterate between an unspecified number of triggers/events, and each of these events can have their own commands. A command is meant to be an action such as showing a text, adding an item, checking an item, etc.
Implemented method of player interaction is done by entering a trigger. This should be easily adapted to interaction by clicking.